requirements.txt
# Core dependencies
curl_cffi>=0.5.0
telethon>=1.28.0
python-dotenv>=1.0.0

# Chat overlay (optional)
websockets>=11.0
cloudscraper>=1.2.0